[config] Add structured logs
Organizing log data in a structured format makes easier
to read and analyze them.

With this commit, log messages will always have a structured
format. The default mode is to print the messages to the
console in a plain format, but it can also be configured
to print them in JSON format.

We have added structured logs to GrimoireLab using the
'structlog' package.

By default, tests will run silent. If developers want to get
log messages printed to the console, activate the environment
variable 'GRIMOIRELAB_TESTING_VERBOSE'.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>